requirements.txt
numpy
sounddevice
faster-whisper
keyboard
pyperclip
pyyaml
